requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
channels==2.4.0
channels-redis==2.4.2
chardet==4.0.0
constantly==15.1.0
cryptography==3.3.1
daphne==2.5.0
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
hiredis==1.1.0
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
msgpack==0.6.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pytz==2020.5
requests==2.25.1
service-identity==18.1.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.12.1
urllib3==1.26.2
wikia==1.4.4
zope.interface==5.2.0
